DHSCFT-422: Move the spine chart table header and rows to separate files
